add packages for test
「babel-core: "7.0.0-bridge.0"」はこのエラーを回避するため

```
Requires Babel "^7.0.0-0", but was loaded with "6.26.3". If you are sure you have a compatible version of @babel/core, it is likely that something in your build process is loading the wrong version. Inspect the stack trace of this error to look for the first entry that doesn't mention "@babel/core" or "babel-core" to see what is calling Babel.
```

[reference]
https://github.com/FormidableLabs/spectacle/issues/522#issuecomment-419599438